+++ src/smros/include/utils.h
@@ -11,12 +11,12 @@ void FilterPointCloud(pcl::PointCloud<SemanticSurfel>::Ptr in,
                       pcl::PointCloud<SemanticSurfel>::Ptr out,
                       float resolution);
 
-Eigen::Matrix4f ComputePoseWithNdt(pcl::PointCloud<SemanticSurfel>::Ptr input_points,
+Eigen::Matrix4d ComputePoseWithNdt(pcl::PointCloud<SemanticSurfel>::Ptr input_points,
                                    pcl::PointCloud<SemanticSurfel>::Ptr target_points,
-                                   Eigen::Matrix4f &initial_pose);
+                                   Eigen::Matrix4d &initial_pose);
 
-Eigen::Matrix4f ComputePoseWithNonlinear(pcl::PointCloud<SemanticSurfel>::Ptr input_points,
+Eigen::Matrix4d ComputePoseWithNonlinear(pcl::PointCloud<SemanticSurfel>::Ptr input_points,
                                          pcl::PointCloud<SemanticSurfel>::Ptr target_points,
-                                         Eigen::Matrix4f &initial_pose);
+                                         Eigen::Matrix4d &initial_pose);
 
 #endif //SMROS_UTILS_H
