+++ src/smros/src/backendopt.cpp
@@ -7,86 +7,86 @@
 using namespace gtsam;
 
 BackEndOpt::BackEndOpt() : pose_graph_(), initials_(), results_() {
-  auto &diag = info_.diagonal();
-  // translational noise is smaller than ro tational noise.
-  double transNoise = 1.0;
-  double rotNoise = 1.0;
-  diag[0] = (transNoise * transNoise);
-  diag[1] = (transNoise * transNoise);
-  diag[2] = (transNoise * transNoise);
+    auto &diag = info_.diagonal();
+    // translational noise is smaller than ro tational noise.
+    double transNoise = 1.0;
+    double rotNoise = 1.0;
+    diag[0] = (transNoise * transNoise);
+    diag[1] = (transNoise * transNoise);
+    diag[2] = (transNoise * transNoise);
 
-  diag[3] = (rotNoise * rotNoise);
-  diag[4] = (rotNoise * rotNoise);
-  diag[5] = (rotNoise * rotNoise);
+    diag[3] = (rotNoise * rotNoise);
+    diag[4] = (rotNoise * rotNoise);
+    diag[5] = (rotNoise * rotNoise);
 }
 
 bool BackEndOpt::addEdge(int from, int to, const Eigen::Matrix4d &measurement,
                          const Eigen::Matrix<double, 6, 6> &information) {
-  auto odometry_model = noiseModel::Gaussian::Information(information);
+    auto odometry_model = noiseModel::Gaussian::Information(information);
 
-  NonlinearFactor::shared_ptr factor(
-      new BetweenFactor<Pose3>(from, to, Pose3(measurement), odometry_model));
-  pose_graph_.add(factor);
-  return false;
+    NonlinearFactor::shared_ptr factor(
+            new BetweenFactor<Pose3>(from, to, Pose3(measurement), odometry_model));
+    pose_graph_.add(factor);
+    return false;
 }
 
 bool BackEndOpt::addEdge(int from, int to, const Eigen::Matrix4d &measurement) {
-  auto odometry_model = noiseModel::Gaussian::Information(info_);
+    auto odometry_model = noiseModel::Gaussian::Information(info_);
 
-  NonlinearFactor::shared_ptr factor(
-      new BetweenFactor<Pose3>(from, to, Pose3(measurement), odometry_model));
-  pose_graph_.add(factor);
-  return false;
+    NonlinearFactor::shared_ptr factor(
+            new BetweenFactor<Pose3>(from, to, Pose3(measurement), odometry_model));
+    pose_graph_.add(factor);
+    return false;
 }
 
 bool BackEndOpt::setInitialValues(int id,
                                   const Eigen::Matrix4d &initial_estimate) {
-  bool addPrior = initials_.empty();
+    bool addPrior = initials_.empty();
 
-  if (!initials_.exists(id))
-    initials_.insert(id, Pose3(initial_estimate));
-  else
-    initials_.update(id, Pose3(initial_estimate));
+    if (!initials_.exists(id))
+        initials_.insert(id, Pose3(initial_estimate));
+    else
+        initials_.update(id, Pose3(initial_estimate));
 
-  // also update intermediate result.
-  if (results_.exists(id))
-    results_.update(id, Pose3(initial_estimate));
-  else
-    results_.insert(id, Pose3(initial_estimate));
+    // also update intermediate result.
+    if (results_.exists(id))
+        results_.update(id, Pose3(initial_estimate));
+    else
+        results_.insert(id, Pose3(initial_estimate));
 
-  if (addPrior) {
-    Vector6 diagonal;
-    diagonal << 1e-6, 1e-6, 1e-6, 1e-6, 1e-6, 1e-6;
-    noiseModel::Diagonal::shared_ptr priorModel =
-        noiseModel::Diagonal::Variances(diagonal);
-    pose_graph_.add(PriorFactor<Pose3>(id, Pose3(), priorModel));
-  }
-  return false;
+    if (addPrior) {
+        Vector6 diagonal;
+        diagonal << 1e-6, 1e-6, 1e-6, 1e-6, 1e-6, 1e-6;
+        noiseModel::Diagonal::shared_ptr priorModel =
+                noiseModel::Diagonal::Variances(diagonal);
+        pose_graph_.add(PriorFactor<Pose3>(id, Pose3(), priorModel));
+    }
+    return false;
 }
 
 bool BackEndOpt::optimize(int num_iters) {
-  LevenbergMarquardtParams params;
-  params.maxIterations = num_iters;
-  params.setVerbosity(
-      "TERMINATION"); // this will show info about stopping conditions
-  params.setVerbosity("SILENT");
-  LevenbergMarquardtOptimizer optimizer(pose_graph_, initials_, params);
+    LevenbergMarquardtParams params;
+    params.maxIterations = num_iters;
+    params.setVerbosity(
+            "TERMINATION"); // this will show info about stopping conditions
+    params.setVerbosity("SILENT");
+    LevenbergMarquardtOptimizer optimizer(pose_graph_, initials_, params);
 
-  results_ = optimizer.optimize();
-  return true;
+    results_ = optimizer.optimize();
+    return true;
 }
 
 bool BackEndOpt::updatePoses(
-    std::vector<Eigen::Matrix4f, Eigen::aligned_allocator<Eigen::Matrix4f>>
+        std::vector<Eigen::Matrix4d, Eigen::aligned_allocator<Eigen::Matrix4d>>
         &pose) {
-  for (const auto &pair : results_) {
-    if (pair.key >= pose.size())
-      continue;
+    for (const auto &pair: results_) {
+        if (pair.key >= pose.size())
+            continue;
 
-    pose[pair.key] = pair.value.cast<Pose3>().matrix().cast<float>();
-  }
+        pose[pair.key] = pair.value.cast<Pose3>().matrix();
+    }
 
-  return false;
+    return false;
 }
 
 BackEndOpt::~BackEndOpt() {}
